+++ CHANGES.rst
@@ -22,7 +22,10 @@ Release History
 2.2.0 (unreleased)
 ==================
 
+**Bug fixes**
 
+- The synapse methods ``filt`` and ``filtfilt`` now support lists as input.
+  (`#1123 <https://github.com/nengo/nengo/pull/1123>`_)
 
 2.1.2 (June 27, 2016)
 =====================